consolidate code into same package
