+++ pkg/scanner/executable.go
@@ -19,59 +19,51 @@ specific language governing permissions and limitations
 under the License.
 */
 
-package app
+package scanner
 
 import (
 	"fmt"
 	"net/http"
 	"os"
 
-	"github.com/blackducksoftware/perceptor-scanner/pkg/scanner"
-
 	"github.com/prometheus/client_golang/prometheus"
 
 	log "github.com/sirupsen/logrus"
 )
 
-// PerceptorScanner handles scanning containers
-type PerceptorScanner struct {
-	scannerManager *scanner.Scanner
-	config         *scanner.Config
-}
-
-// NewPerceptorScanner creates a new PerceptorScanner object
-func NewPerceptorScanner(configPath string) (*PerceptorScanner, error) {
-	config, err := scanner.GetConfig(configPath)
+// RunPerceptorScanner runs perceptor-scanner
+func RunPerceptorScanner(configPath string, stop <-chan struct{}) {
+	config, err := GetConfig(configPath)
 	if err != nil {
-		return nil, fmt.Errorf("Failed to load configuration: %v", err)
+		log.Error(err.Error())
+		panic(err)
 	}
 
 	level, err := config.GetLogLevel()
 	if err != nil {
-		return nil, err
+		log.Error(err.Error())
+		panic(err)
 	}
 	log.SetLevel(level)
 
-	sm, err := scanner.NewScanner(config)
+	scannerManager, err := NewScannerManager(config)
 	if err != nil {
-		return nil, fmt.Errorf("unable to instantiate scanner: %v", err)
+		log.Error(err.Error())
+		panic(err)
 	}
 
-	return &PerceptorScanner{scannerManager: sm, config: config}, nil
-}
-
-// Run starts the PerceptorScanner looking for scan jobs
-func (ps *PerceptorScanner) Run(stopCh <-chan struct{}) {
 	prometheus.Unregister(prometheus.NewProcessCollector(os.Getpid(), ""))
 	prometheus.Unregister(prometheus.NewGoCollector())
 
-	ps.scannerManager.StartRequestingScanJobs()
+	scannerManager.StartRequestingScanJobs()
 
 	http.Handle("/metrics", prometheus.Handler())
 
-	addr := fmt.Sprintf(":%d", ps.config.Port)
-	log.Infof("successfully instantiated scanner %+v, serving on %s", ps.scannerManager, addr)
-	http.ListenAndServe(addr, nil)
+	addr := fmt.Sprintf(":%d", config.Port)
+	log.Infof("successfully instantiated scanner %+v, serving on %s", scannerManager, addr)
+	go func() {
+		http.ListenAndServe(addr, nil)
+	}()
 
-	<-stopCh
+	<-stop
 }
